LF-107: Replace old nxtp implementation with new one
